Made Slight Changes
Quest Mark:
Slowed the animation for it

UI Quest:
Moved it to the side screen to fit the game.
Added the Ashbringer Music to it.

The Portal:
Added a hidden button to the top of the play some music.
Still didn't find a method to change the color of the portal itself.